ML/classify/requirements.txt
paho-mqtt==1.6.1
tensorflow==2.10.1
scikit-learn==1.6.0
python-dotenv==1.0.0